Move incident and patient presenter to middleware.
