+++ cdx_core/app/views/patients/index.html.haml
@@ -16,7 +16,7 @@
             = image_tag('img/user.png')
             %h3
               = I18n.t('patients.index.patient', count: @total)
-          = react_component "PatientsIndexTable", patients: Presenters::Patients.index_table(@patients),
+          = react_component "PatientsIndexTable", patients: Patients::Presenter.index_table(@patients),
               orderBy: params["order_by"] || "patients.name", searchTerm: params['address']
           .pagination
             = render 'shared/pagination'
